Initial revision
This is a new SAT solver written from scratch. It aims to be efficient
on multicore architectures.

Most things are still missing; the watchlist scheme has been
implemented, but there is no support for even basic things like clause
learning. Whenever a conflict is encountered, the solver simply
backtracks to a random decision level.

This code is licensed under GPL3; see the file COPYING for the complete
license.

Signed-off-by: Vegard Nossum <vegardno@ifi.uio.no>