Pass watch to solver::find_new_watch
Instead of trying to fix a clause's watched literals in general, we now
pass the watch that was just falsified as an argument to the function.
If there is no other unassigned literal in the clause, we always imply
the other watched literal (propagation) and we let solver::implication
take care of whether there was a conflict or not.

Signed-off-by: Vegard Nossum <vegardno@ifi.uio.no>