Add restart strategies "geometric" and "nested"
The geometric restart strategy allows a certain number of decisions,
after which this the number is increased by a certain fraction.

The nested restart strategy allows one to specify inner and outer
restart strategies; when the inner strategy indicates a restart, the
outer strategy is checked as well. If the outer strategy indicates a
restart, the inner strategy is reset.

The idea for nested restarts came from:

A. Biere. Adaptive Restart Control for Conflict Driven SAT Solvers. In
Proc. 11th Intl. Conf. on Theory and Applications of Satisfiability
Testing (SAT'08), Lecture Notes in Computer Science (LNCS) vol. 4996,
Springer 2008.

Signed-off-by: Vegard Nossum <vegardno@ifi.uio.no>