Add more restart strategies: restart_not, restart_and, restart_or
It is doubtful whether these will all be really useful. But as it turns
out that restart strategies are highly composeable. For example, if you
use restart_and<restart_fixed<100>, restart_luby>, you'll get sequences
like: 100, 100, 200, 100, 100, 200, 400, 100, ... (i.e. each term in the
luby sequence is multiplied by 100).

It could be useful in the future to change the restart interface to
return a number (which is the number of decisions until the next restart)
instead of returning just true or false (which is whether we want to
restart now or not), simply because this is more efficient.

Signed-off-by: Vegard Nossum <vegardno@ifi.uio.no>