Fix bug in analyze_1uip
The solver would abort with this error:

solver: clause.hh:127: literal& clause::operator[](unsigned int): Assertion `i < data->size' failed.
Aborted

It turns out that we introduced a bug in:

commit 70b658c173740226e4904a4cc31a5084bb161552
Author: Vegard Nossum <vegardno@ifi.uio.no>
Date:   Mon Nov 21 12:46:22 2011 +0100

    Move analyze_1uip into its own file

where we would dereference the pointer after incrementing it instead of
dereference the pointer and increment the atomic value it pointed to.

Signed-off-by: Vegard Nossum <vegardno@ifi.uio.no>