Allow threads to allocate new clauses independently of each other
This patch is rather intrusive, but it is necessary.

Each clause is now identified by the owning thread and its index for that
thread (instead of just a global index).

This means that each thread now needs to have N watchlists if there are N
threads running.

The upside of all this is that we don't need the atomic increment of the
global clause id/counter when creating new clauses. Moreover, we can also
detach clauses without requiring an atomic access.

Signed-off-by: Vegard Nossum <vegardno@ifi.uio.no>