Conflict analysis should end in an implication, not a decision
Specifically, if we make it a decision, the solver may return to the
exact same decision level again and learn the exact same clause again
because it doesn't have the learnt clause as the reason why the
literal was forced. If it does, it will return to an earlier decision
level because the reason clause will be taken into account in the
resolution process.

Signed-off-by: Vegard Nossum <vegardno@ifi.uio.no>