Fix clause attachment to possibly propagate or return unsat
The problem is that we may receive clauses from other threads while
we have a different partial assignment (due to discovered unit
clauses), and these clauses may be propagating or even falsified as
we try to attach them.

We now detect these cases in attach() and provide other functions
for the cases where we know the outcome of attaching it (e.g. for
asserting learnt clauses).

Signed-off-by: Vegard Nossum <vegardno@ifi.uio.no>