Add support for selectively accepting received clauses
For now, since we only have a very simple clause sending filter that all
threads adhere to, we don't need to do anything at all in the new clause
receival filter.

However, when threads have different parameters, it could be interesting
to have a receive plugin that does something more fancy (like determining
whether the clause in question is relevant for the current thread).

Signed-off-by: Vegard Nossum <vegardno@ifi.uio.no>