Add some prefetch instructions to propagate_watchlists
I actually measured what prefetch distances (i.e. offset into the
future) that were optimal on my laptop, and these were the values.

This is not to say that these are the optimal values for all machines,
and certainly not after we possibly change the source code, so we
should beware of that.

But for now, this gives a nice speedup of nearly 25% for certain
benchmarks.

Signed-off-by: Vegard Nossum <vegardno@ifi.uio.no>