Update the VSIDS heuristic
This makes it a lot more like the one in minisat. We keep track only
of variables, not literals, and use polarity caching to determine the
polarity of the decision literal.

We still don't use the same scoring technique, though, because we
don't actually decay the activities. So this is more like VSIS...

Signed-off-by: Vegard Nossum <vegardno@ifi.uio.no>