Make print_stdio a generic plugin
Also fix some bugs in plugin_list that had to be fixed for this to
work properly.

Unfortunately, I don't know of a way to call the plugin constructors
of the individual elements of the plugin list with the solver as the
argument, so instead we provide a start() function that also takes a
reference to the solver.

Signed-off-by: Vegard Nossum <vegardno@ifi.uio.no>