Implement a basic clause redundancy check for received clauses
Clauses that we received from another thread may already have been
discovered by the current thread, so attaching it twice is simply a
waste of work (it slows down propagation for no purpose).

This still needs some work, because it can potentially shorten
clauses that were received; however, we don't yet have the necessary
infrastructure for announcing that a clause should be replaced.

We nevertheless observe a 20% speedup for a certain benchmark and 4
threads.

Signed-off-by: Vegard Nossum <vegardno@ifi.uio.no>