Add our own definition of assert() that aborts on failure
The libc/stdlib definition of assert() seems to call exit(EXIT_FAILURE)
or so when the assertion fails. This is WRONG, since we actually want
to obtain the core dump, stack trace, etc., all of which are gone if
exit() is called.

So we code our own version of it.

Signed-off-by: Vegard Nossum <vegardno@ifi.uio.no>