Handle attachment of falsified clauses
We previously didn't handle this case because there was no way we would
even attempt to attach a falsified clause (the only place where we
would attach clauses was from conflict analysis, and analyze_1uip never
attaches a falsified clause anyway). With clause sharing between
threads, this was now possible, but rare. Fix it.

Signed-off-by: Vegard Nossum <vegardno@ifi.uio.no>