Add support for other clause types
This is a pretty big change. We essentially replace clauses with an
abstract "reason" class whenever we're dealing with reasons for
propagating a certain variable.

The idea here is that since we don't know what clause types we need to
support in advance (since the user might want to add propagation
engines for other clause types), the user needs to supply the reason
class. The reason class should be a kind of a union class that
represents at any time exactly one type of clause.

Other kinds of constraints can always be modelled using a conjunction
of regular (disjunctive) clauses, and when a propagation happens, the
propagation can always be modelled using exactly a regular clause.

We could use some kind of runtime polymorphism for this and not require
the user to implement the reason class at all. However, there doesn't
seem to be any way to do this efficiently (without runtime overhead).

Signed-off-by: Vegard Nossum <vegardno@ifi.uio.no>